Return time logged in hours for overview

+++ server/Timelogger.Api/Models/ProjectOverviewModel.cs
@@ -6,7 +6,7 @@ namespace Timelogger.Api
   {
     public int Id { get; set; }
     public string Name { get; set; }
-    public int TimeLogged { get; set; }
+    public double TimeLogged { get; set; }
     public DateTime Deadline { get; set; }
   }
 }
